agent/backend/requirements.txt
# Core dependencies
fastapi==0.124.4
uvicorn==0.38.0
pydantic==2.12.5
pydantic-settings==2.12.0
starlette==0.50.0
# Claude SDK
claude-agent-sdk==0.1.19
# Database
psycopg2-binary==2.9.9
# Utilities
python-multipart==0.0.20
httpx==0.27.0
PyJWT==2.8.0
apscheduler==3.10.4
# Browser automation
playwright==1.57.0
# Redis caching
redis==5.0.1
